- Restructured code, main logic is now in tx-serializer.js
